test: update create sever name
it's needed that it has different names bc it's in the same package
