+++ Makefile
@@ -1,4 +1,7 @@
-coverage:
-	go test -v ./... -covermode=atomic -coverpkg=./... -count=1  -race -timeout=30m -coverprofile=coverage.out && go tool cover -html=coverage.out
+test:
+	go test -v ./... -covermode=atomic -coverpkg=./... -count=1  -race -timeout=30m -coverprofile=coverage.out
 
-.PRONY: coverage
+coverage:
+	go tool cover -html=coverage.out
+
+.PRONY: test coverage
