+++ cmd/server/handler/index_test.go
@@ -1,0 +1,18 @@
+package controllers_test
+
+import (
+	"bytes"
+	"net/http"
+	"net/http/httptest"
+)
+
+func createRequestTest(
+	method string,
+	url string,
+	body string,
+) (*http.Request, *httptest.ResponseRecorder) {
+	req := httptest.NewRequest(method, url, bytes.NewBuffer([]byte(body)))
+	req.Header.Add("Content-Type", "application/json")
+
+	return req, httptest.NewRecorder()
+}
