+++ go.mod
@@ -4,28 +4,14 @@ go 1.18
 
 require (
 	github.com/gin-gonic/gin v1.8.1
-
 	github.com/stretchr/testify v1.7.2
 
-	github.com/go-playground/assert/v2 v2.0.1
-	github.com/stretchr/testify v1.7.1
-
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/stretchr/objx v0.1.0 // indirect
-
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/stretchr/objx v0.1.0 // indirect
-	github.com/stretchr/testify v1.7.2 // indirect
-
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
